refactor(jans-cedarling): replace old PolicyStore with updated implementation
