+++ jans-cedarling/cedarling/src/common/policy_store/yaml_store.rs
@@ -1,67 +1,36 @@
 use super::{
-    super::{
-        super::cedar_schema::CedarSchemaJson, parse_maybe_cedar_version, parse_option_string,
-        trusted_issuer_metadata::TrustedIssuerMetadata, CedarSchema,
-    },
-    PolicyContent, PolicyStoreSource,
+    parse_maybe_cedar_version, parse_option_string, trusted_issuer_metadata::TrustedIssuerMetadata,
+    CedarSchema, PolicyContent, PolicyStore,
 };
-use base64::prelude::*;
-use cedar_policy::{Policy, PolicyId, Schema};
+use crate::common::cedar_schema::CedarSchemaJson;
+use cedar_policy::{
+    ffi::{self},
+    Policy, PolicyId, PolicySet, Schema,
+};
 use semver::Version;
 use serde::{de, Deserialize, Deserializer};
-use std::collections::HashMap;
+use std::{collections::HashMap, str::FromStr};
 
-/// Policy Store generated by Agama Lab.
+/// Policy Store used for internal testing which supports readable
+/// Cedar policies and schema.
 #[derive(Deserialize)]
-pub struct PolicyStoreSourceJson {
-    #[serde(deserialize_with = "parse_maybe_cedar_version")]
-    cedar_version: Option<Version>,
-    policy_stores: HashMap<String, PolicyStoreDataJson>,
-}
-
-impl TryFrom<PolicyStoreSourceJson> for PolicyStoreSource {
-    type Error = LoadFromJsonError;
-
-    fn try_from(json_store: PolicyStoreSourceJson) -> Result<Self, Self::Error> {
-        match json_store.policy_stores.len() {
-            0 => Err(LoadFromJsonError::NoPolicyStore),
-            1 => {
-                let (_id, store) = json_store
-                    .policy_stores
-                    .into_iter()
-                    .next()
-                    .expect("Should have one policy store");
-                Ok(Self {
-                    name: store.name,
-                    description: store.description,
-                    cedar_version: json_store.cedar_version,
-                    policies: store.policies,
-                    cedar_schema: store.cedar_schema,
-                    trusted_issuers: store.trusted_issuers,
-                })
-            },
-            _ => Err(LoadFromJsonError::MoreThanOnePolicyStore),
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Deserialize)]
-#[allow(dead_code)]
-pub struct PolicyStoreDataJson {
+pub struct PolicyStoreYaml {
     #[serde(deserialize_with = "parse_option_string", default)]
     pub name: Option<String>,
     #[serde(deserialize_with = "parse_option_string", default)]
     pub description: Option<String>,
-    #[serde(deserialize_with = "parse_encoded_policies")]
+    #[serde(deserialize_with = "parse_maybe_cedar_version", default)]
+    pub cedar_version: Option<Version>,
+    #[serde(deserialize_with = "parse_human_readable_policies", default)]
     pub policies: HashMap<String, PolicyContent>,
-    #[serde(alias = "schema", deserialize_with = "parse_encoded_schema")]
+    #[serde(alias = "schema", deserialize_with = "parse_human_readable_schema")]
     pub cedar_schema: CedarSchema,
     #[serde(default)]
     pub trusted_issuers: HashMap<String, TrustedIssuerMetadata>,
 }
 
-// Parse a Hashmap of Base64 encoded Cedar policies
-fn parse_encoded_policies<'de, D>(
+// Parse a Hashmap of Human-redable Cedar policies
+fn parse_human_readable_policies<'de, D>(
     deserializer: D,
 ) -> Result<HashMap<String, PolicyContent>, D::Error>
 where
@@ -71,9 +40,8 @@ where
     struct RawPolicy {
         description: String,
         creation_date: String,
-        /// Base64 encoded JSON Cedar Policy
         #[serde(rename = "policy_content")]
-        encoded_policy: String,
+        cedar_policy: String,
     }
 
     let raw_policies = <HashMap<String, RawPolicy> as de::Deserialize>::deserialize(deserializer)?;
@@ -81,8 +49,7 @@ where
 
     for (policy_id, policy) in raw_policies {
         // parse the decoded policy string into a cedar_policy::Policy struct
-        let decoded_policy = decode_b64_string::<D>(policy.encoded_policy)?;
-        let cedar_policy = Policy::parse(Some(PolicyId::new(&policy_id)), decoded_policy)
+        let cedar_policy = Policy::parse(Some(PolicyId::new(&policy_id)), policy.cedar_policy)
             .map_err(de::Error::custom)?;
 
         let agama_policy = PolicyContent {
@@ -97,51 +64,72 @@ where
     Ok(policies)
 }
 
-// Parse a Base64 encoded Cedar schema
-fn parse_encoded_schema<'de, D>(deserializer: D) -> Result<CedarSchema, D::Error>
+// Parse a human-readable Cedar schema
+fn parse_human_readable_schema<'de, D>(deserializer: D) -> Result<CedarSchema, D::Error>
 where
     D: Deserializer<'de>,
 {
-    let encoded_schema = String::deserialize(deserializer)?;
-    let decoded_schema = decode_b64_string::<D>(encoded_schema)?;
-    let schema = Schema::from_json_str(&decoded_schema)
+    let cedar_schema = String::deserialize(deserializer)?;
+    let schema = Schema::from_str(&cedar_schema)
         .map_err(|e| de::Error::custom(format!("Failed to load Schema from JSON: {:?}", e)))?;
-    let json = serde_json::from_str::<CedarSchemaJson>(&decoded_schema).map_err(|e| {
+
+    // we convert the Schema to a JSON format because it's needed for the
+    // rest of the code.
+    let ffi_schema = ffi::Schema::Cedar(cedar_schema.clone());
+    let ffi_json_schema = match ffi::schema_to_json(ffi_schema) {
+        ffi::SchemaToJsonAnswer::Success { json, .. } => json,
+        ffi::SchemaToJsonAnswer::Failure { errors } => {
+            return Err(de::Error::custom(format!(
+                "Error converting Cedar schema to JSON: {:?}",
+                errors
+            )))
+        },
+    };
+    let json = serde_json::to_string(&ffi_json_schema).unwrap();
+    let json = serde_json::from_str::<CedarSchemaJson>(&json).map_err(|e| {
         de::Error::custom(format!("Failed to Load CedarSchemaJson from JSON: {:?}", e))
     })?;
+
     Ok(CedarSchema { schema, json })
 }
 
-fn decode_b64_string<'de, D>(value: String) -> Result<String, D::Error>
-where
-    D: de::Deserializer<'de>,
-{
-    let buf = BASE64_STANDARD
-        .decode(value)
-        .map_err(|e| de::Error::custom(format!("Failed to decode Base64 encoded string: {}", e)))?;
-    String::from_utf8(buf)
-        .map_err(|e| de::Error::custom(format!("Failed to decode Base64 encoded string: {}", e)))
+impl From<PolicyStoreYaml> for PolicyStore {
+    fn from(yaml_store: PolicyStoreYaml) -> Self {
+        let mut policy_set = PolicySet::new();
+        for (_id, policy) in &yaml_store.policies {
+            // this expect should never be called because we defined our
+            // own policies when we parsed the policy store and they
+            // never are template-linked.
+            policy_set
+                .add(policy.policy_content.clone())
+                .expect("Expected a non-template linked policy");
+        }
+
+        Self {
+            name: yaml_store.name,
+            description: yaml_store.description,
+            cedar_version: yaml_store.cedar_version,
+            policies: yaml_store.policies,
+            cedar_schema: yaml_store.cedar_schema,
+            trusted_issuers: yaml_store.trusted_issuers,
+            policy_set,
+        }
+    }
 }
 
 #[derive(Debug, thiserror::Error)]
-pub enum LoadFromJsonError {
-    #[error("The JSON policy stores does not contain a single policy store.")]
-    NoPolicyStore,
-    #[error("The JSON policy stores contain more than one policy store.")]
-    MoreThanOnePolicyStore,
-    #[error("Failed to load policy store from JSON: {0}")]
-    Deserialization(serde_json::Error),
+pub enum LoadFromYamlError {
+    #[error("Failed to load policy store from YAML: {0}")]
+    Deserialization(serde_yml::Error),
 }
 
 #[cfg(test)]
 mod test {
-    use super::super::super::super::{cedar_schema::CedarSchemaJson, policy_store::CedarSchema};
+    use super::super::super::{cedar_schema::CedarSchemaJson, policy_store::CedarSchema};
     use super::super::PolicyContent;
-    use super::{parse_encoded_policies, parse_encoded_schema};
-    use base64::prelude::*;
+    use super::{parse_human_readable_policies, parse_human_readable_schema};
     use cedar_policy::{ffi, Policy, PolicyId, Schema};
     use serde::Deserialize;
-    use serde_json::json;
     use std::collections::HashMap;
     use std::str::FromStr;
     use test_utils::assert_eq;
@@ -165,10 +153,10 @@ mod test {
     }
 
     #[test]
-    fn can_parse_encoded_policies() {
+    fn can_parse_human_readable_policies() {
         #[derive(Deserialize, Debug, PartialEq)]
         struct TestPolicy {
-            #[serde(deserialize_with = "parse_encoded_policies")]
+            #[serde(deserialize_with = "parse_human_readable_policies")]
             pub policies: HashMap<String, PolicyContent>,
         }
 
@@ -183,13 +171,11 @@ mod test {
                 policy_content: Policy::parse(
                     Some(PolicyId::new(policy_id)),
                     r#"@id("Admin")
-permit
-(
+permit (
  principal == somecompany::store::Role::"Admin",
  action in [somecompany::store::Action::"DELETE",somecompany::store::Action::"GET",somecompany::store::Action::"PUT"],
  resource == somecompany::store::HTTP_Request::"root"
-)
-;"#.to_string(),
+);"#.to_string(),
                 )
                 .expect("should parse cedar policy"),
             },
@@ -203,13 +189,11 @@ permit
                 policy_content: Policy::parse(
                     Some(PolicyId::new(policy_id)),
                     r#"@id("Member")
-permit
-(
+permit (
  principal == somecompany::store::Role::"Member",
  action in [somecompany::store::Action::"PUT"],
  resource == somecompany::store::HTTP_Request::"root"
-)
-;"#
+);"#
                     .to_string(),
                 )
                 .expect("should parse cedar policy"),
@@ -217,23 +201,30 @@ permit
         );
         let expected = TestPolicy { policies };
 
-        let json_policies = json!({
-            "policies": {
-                "some_policy_id": {
-                    "description": "Admin",
-                    "creation_date": "2024-11-07T07:49:11.813002",
-                    "policy_content": "QGlkKCJBZG1pbiIpCnBlcm1pdAooCiBwcmluY2lwYWwgPT0gc29tZWNvbXBhbnk6OnN0b3JlOjpSb2xlOjoiQWRtaW4iLAogYWN0aW9uIGluIFtzb21lY29tcGFueTo6c3RvcmU6OkFjdGlvbjo6IkRFTEVURSIsc29tZWNvbXBhbnk6OnN0b3JlOjpBY3Rpb246OiJHRVQiLHNvbWVjb21wYW55OjpzdG9yZTo6QWN0aW9uOjoiUFVUIl0sCiByZXNvdXJjZSA9PSBzb21lY29tcGFueTo6c3RvcmU6OkhUVFBfUmVxdWVzdDo6InJvb3QiCikKOw=="
-                },
-                "another_policy_id": {
-                    "description": "Member",
-                    "creation_date": "2024-11-07T07:50:05.520757",
-                    "policy_content": "QGlkKCJNZW1iZXIiKQpwZXJtaXQKKAogcHJpbmNpcGFsID09IHNvbWVjb21wYW55OjpzdG9yZTo6Um9sZTo6Ik1lbWJlciIsCiBhY3Rpb24gaW4gW3NvbWVjb21wYW55OjpzdG9yZTo6QWN0aW9uOjoiUFVUIl0sCiByZXNvdXJjZSA9PSBzb21lY29tcGFueTo6c3RvcmU6OkhUVFBfUmVxdWVzdDo6InJvb3QiCikKOw=="
-                }
-            }
-        });
+        let yaml_policies = r#"policies:
+          some_policy_id:
+            description: Admin
+            creation_date: 2024-11-07T07:49:11.813002
+            policy_content: |-
+              @id("Admin")
+              permit (
+               principal == somecompany::store::Role::"Admin",
+               action in [somecompany::store::Action::"DELETE",somecompany::store::Action::"GET",somecompany::store::Action::"PUT"],
+               resource == somecompany::store::HTTP_Request::"root"
+              );
+          another_policy_id:
+            description: Member
+            creation_date: 2024-11-07T07:50:05.520757
+            policy_content: |-
+              @id("Member")
+              permit (
+               principal == somecompany::store::Role::"Member",
+               action in [somecompany::store::Action::"PUT"],
+               resource == somecompany::store::HTTP_Request::"root"
+              );"#;
 
-        let parsed_policies = serde_json::from_value::<TestPolicy>(json_policies)
-            .expect("Failed to parse policies from JSON");
+        let parsed_policies = serde_yml::from_str::<TestPolicy>(yaml_policies)
+            .expect("Failed to parse policies from YAML");
 
         assert_eq!(
             parsed_policies, expected,
@@ -242,10 +233,10 @@ permit
     }
 
     #[test]
-    fn can_parse_encoded_schema() {
+    fn can_parse_human_readable_schema() {
         #[derive(Debug, PartialEq, Deserialize)]
         struct TestPolicyStore {
-            #[serde(alias = "schema", deserialize_with = "parse_encoded_schema")]
+            #[serde(alias = "schema", deserialize_with = "parse_human_readable_schema")]
             pub cedar_schema: CedarSchema,
         }
 
@@ -255,30 +246,49 @@ permit
     entity Issue = {"country": String, "org_id": String};
     entity Role;
     entity TrustedIssuer = {"issuer_entity_id": Url};
+
     entity User in [Role] = {"country": String, "email": String, "sub": String, "username": String};
     entity Workload = {"client_id": String, "iss": TrustedIssuer, "name": String, "org_id": String};
+
     entity id_token = {"acr": String, "amr": String, "aud": String, "exp": Long, "iat": Long, "iss": TrustedIssuer, "jti": String, "sub": String};
     action "Update" appliesTo {
+
         principal: [Workload, User, Role],
         resource: [Issue],
+
         context: {}
     };
 }"#;
 
-        let (cedar_schema, json_schema) = load_from_cedar_schema_from_str(cedar_schema_str);
+        let (cedar_schema, _) = load_from_cedar_schema_from_str(cedar_schema_str);
         let expected = TestPolicyStore { cedar_schema };
 
-        let encoded_schema = BASE64_STANDARD.encode(json_schema.as_bytes());
-        let json_schema = json!({
-            "schema": encoded_schema,
-        });
+        let yaml_schema = r#"schema: |-
+    namespace Jans {
+        type Url = {"host": String, "path": String, "protocol": String};
 
-        let parsed = serde_json::from_value::<TestPolicyStore>(json_schema)
-            .expect("Failed to parse schema from JSON");
+        entity Access_token = {"aud": String, "exp": Long, "iat": Long, "iss": TrustedIssuer, "jti": String};
+        entity Issue = {"country": String, "org_id": String};
+        entity Role;
+        entity TrustedIssuer = {"issuer_entity_id": Url};
+        entity User in [Role] = {"country": String, "email": String, "sub": String, "username": String};
+        entity Workload = {"client_id": String, "iss": TrustedIssuer, "name": String, "org_id": String};
+        entity id_token = {"acr": String, "amr": String, "aud": String, "exp": Long, "iat": Long, "iss": TrustedIssuer, "jti": String, "sub": String};
+        action "Update" appliesTo {
+            principal: [Workload, User, Role],
+
+            resource: [Issue],
+            context: {}
+
+        };
+    }"#;
+
+        let parsed = serde_yml::from_str::<TestPolicyStore>(yaml_schema)
+            .expect("Failed to parse schema from YAML");
 
         assert_eq!(
             parsed, expected,
-            "Expected to parse encoded schema from JSON correctly"
+            "Expected to parse schema from YAML correctly"
         );
     }
 }
