+++ jans-cedarling/cedarling/src/jwt/test/with_validation/access_token.rs
@@ -21,7 +21,7 @@
 //!   - Tests for errors when the `nbf` has not passed yet.
 
 use super::super::*;
-use crate::common::policy_store::TrustedIssuer;
+use crate::common::policy_store::TrustedIssuerMetadata;
 use crate::jwt::decoding_strategy::JwtDecodingError;
 use crate::jwt::{self, HttpClient, JwtService, TrustedIssuerAndOpenIdConfig};
 use jsonwebtoken::Algorithm;
@@ -135,14 +135,14 @@ fn test_missing_claim(missing_claim: &'static str) {
         .create();
 
     let trusted_idp = TrustedIssuerAndOpenIdConfig::fetch(
-        TrustedIssuer {
+        TrustedIssuerMetadata {
             name: "some_idp".to_string(),
             description: "some_desc".to_string(),
             openid_configuration_endpoint: format!(
                 "{}/.well-known/openid-configuration",
                 server.url()
             ),
-            token_metadata: None,
+            ..Default::default()
         },
         &HttpClient::new().expect("should create http client"),
     )
@@ -285,14 +285,14 @@ fn errors_on_invalid_signature() {
         .create();
 
     let trusted_idp = TrustedIssuerAndOpenIdConfig::fetch(
-        TrustedIssuer {
+        TrustedIssuerMetadata {
             name: "some_idp".to_string(),
             description: "some_desc".to_string(),
             openid_configuration_endpoint: format!(
                 "{}/.well-known/openid-configuration",
                 server.url()
             ),
-            token_metadata: None,
+            ..Default::default()
         },
         &HttpClient::new().expect("should create http client"),
     )
@@ -403,14 +403,14 @@ fn errors_on_expired_token() {
         .create();
 
     let trusted_idp = TrustedIssuerAndOpenIdConfig::fetch(
-        TrustedIssuer {
+        TrustedIssuerMetadata {
             name: "some_idp".to_string(),
             description: "some_desc".to_string(),
             openid_configuration_endpoint: format!(
                 "{}/.well-known/openid-configuration",
                 server.url()
             ),
-            token_metadata: None,
+            ..Default::default()
         },
         &HttpClient::new().expect("should create http client"),
     )
@@ -523,14 +523,14 @@ fn errors_on_token_used_before_nbf() {
         .create();
 
     let trusted_idp = TrustedIssuerAndOpenIdConfig::fetch(
-        TrustedIssuer {
+        TrustedIssuerMetadata {
             name: "some_idp".to_string(),
             description: "some_desc".to_string(),
             openid_configuration_endpoint: format!(
                 "{}/.well-known/openid-configuration",
                 server.url()
             ),
-            token_metadata: None,
+            ..Default::default()
         },
         &HttpClient::new().expect("should create http client"),
     )
